Slidebars updated, search module implemented, actual room name
Slidebars: In the content part now, there are are slider bars too. You
can change position values with them.

Search module: It was hard to find a specific node if you have
thousends. Now node list is in ABC order, and there is a search bar at
the top of the UI. If you would like to find a node, you just need to
write a part of it's name, and the app will list just the maches.

Actual room name: If you select a room, at the settings part, at the top
of the UI, there is a label with it's name.